homolog

+++ src/main/java/com/example/prision/modules/repository/SaleRepository.java
@@ -3,5 +3,5 @@ package com.example.prision.modules.repository;
 import com.example.prision.modules.model.Sale;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SaleRepository extends JpaRepository<Sale, Integer> {
+public interface SaleRepository extends JpaRepository<Sale, Long> {
 }
